+++ src/cmd_ovlpr/mod.rs
@@ -2,3 +2,4 @@
 
 pub mod covered;
 pub mod paf2ovlp;
+pub mod replace;